LogSpy Update
I want to figure out how to render the listbox and when it detects text that it needs to colour, it just colours it.
This would save huge computation time as the program for now goes through the entire log file and colours it in before handing control back to the user.